Speeding up the checkpoint loader using shared memory for copying (#89)
Previously, the checkpoint loader used ptrace POKE_TEXT API to copy the data from the tracer to tracee, which was too slow because POKE_TEXT involves a system call for copying only 8 bytes of data. This commit changes the copying mechanism by allocating a shared memory buffer for the tracer and tracee and injecting code in the tracee to do the copy with much fewer system calls.