Fix fa-cache bug
This line fixes the fa-dcache bug. Before, we were inserting into the (fake) fully
associative cache even when it shouldn't have been in the normal cache, leading to many of
the misses being counted as conflict misses rather than capacity misses. Now we only insert
if it should be inserted, which fixes it.